split client.go into delegations.go
